+++ src/boot/i18n.ts
@@ -22,8 +22,10 @@ declare module 'vue-i18n' {
 /* eslint-enable @typescript-eslint/no-empty-interface */
 
 export default boot(({ app }) => {
+  
   const i18n = createI18n({
-    locale: 'en-US',
+    fallbackLocale: 'en-US',
+    globalInjection: true,
     legacy: false,
     messages,
   });