+++ src/router/routes.ts
@@ -5,8 +5,34 @@ const routes: RouteRecordRaw[] = [
     path: '/',
     component: () => import('layouts/MainLayout.vue'),
     children: [
-      { path: '', component: () => import('pages/IndexPage.vue') },
-      { path: '/setting', component: () => import('pages/SettingPage.vue') }
+      { 
+        path: '', 
+        component: () => import('pages/IndexPage.vue'), 
+        meta: {
+          requiresAuth: true
+        },
+      },
+      { 
+        path: '/setting', 
+        component: () => import('pages/SettingPage.vue') 
+      }
+    ],
+  },
+  {
+    path: '/auth',
+    component: () => import('layouts/AuthLayout.vue'),
+    meta: {
+      redirectAuth: true
+    },
+    children: [
+      { 
+        path: 'login', 
+        component: () => import('pages/auth/SignIn.vue')
+      },
+      { 
+        path: 'register', 
+        component: () => import('pages/auth/SignUp.vue')
+      }
     ],
   },